+++ .ci-scripts/x86-64-unknown-linux-musl-release.bash
@@ -19,7 +19,7 @@ OS=linux-musl
 TRIPLE=${ARCH}-${VENDOR}-${OS}
 
 # Build parameters
-MAKE_PARALLELISM=4
+MAKE_PARALLELISM=8
 BUILD_PREFIX=$(mktemp -d)
 DESTINATION=${BUILD_PREFIX}/lib/pony
 
@@ -30,7 +30,7 @@ PACKAGE=ponyc-${TRIPLE}
 # Cloudsmith configuration
 CLOUDSMITH_VERSION=$(cat VERSION)
 ASSET_OWNER=ponylang
-ASSET_REPO=nightlies
+ASSET_REPO=releases
 ASSET_PATH=${ASSET_OWNER}/${ASSET_REPO}
 ASSET_FILE=${PACKAGE_DIR}/${PACKAGE}.tar.gz
 ASSET_SUMMARY="Pony compiler"