+++ lib/gbenchmark/test/state_assembly_test.cc
@@ -4,11 +4,13 @@
 #pragma clang diagnostic ignored "-Wreturn-type"
 #endif
 
+// clang-format off
 extern "C" {
   extern int ExternInt;
   benchmark::State& GetState();
   void Fn();
 }
+// clang-format on
 
 using benchmark::State;
 
@@ -23,7 +25,7 @@ extern "C" int test_for_auto_loop() {
   for (auto _ : S) {
     // CHECK: .L[[LOOP_HEAD:[a-zA-Z0-9_]+]]:
     // CHECK-GNU-NEXT: subq $1, %rbx
-    // CHECK-CLANG-NEXT: {{(addq \$1,|incq)}} %rax
+    // CHECK-CLANG-NEXT: {{(addq \$1, %rax|incq %rax|addq \$-1, %rbx)}}
     // CHECK-NEXT: jne .L[[LOOP_HEAD]]
     benchmark::DoNotOptimize(x);
   }